Reduce versions of Solidus we test against
This gem doesn't actually work with v1.3 or master. Changes to some of
the admin structure break the admin interfaces for viewing and redeeming
gift cards.

Until we've got a version of this extension that actually works with
those versions of Solidus it's best not to make it seem like this
extension will work without issue.